updating caches an OMMulDiv
